Split the code base up into packages for easier management
